Combined Listener and RunnableThread into one file.
